Implementing cancel registration. Implementing changing registration
