+++ src/app/shared/cranix-ptm-view/cranix-ptm-view.component.html
@@ -108,11 +108,17 @@
     <ng-template>
         <ion-header color="primary">
             <ion-toolbar color="primary">
-                <ion-title>{{'Register room for'|translate}} {{ptmTeacherInRoom.teacher.surName}}, {{ptmTeacherInRoom.teacher.givenName}} </ion-title>
+                <ion-title>{{'Register room for'|translate}} {{ptmTeacherInRoom.teacher.surName}},
+                    {{ptmTeacherInRoom.teacher.givenName}} </ion-title>
                 <ion-buttons slot="end">
                     <ion-button class="ion-no-margin" (click)="doRegisterRoom()">
                         <ion-icon slot="icon-only" color="success" name="checkmark-sharp"></ion-icon>
                     </ion-button>
+                    @if(ptmTeacherInRoom.id != 0) {
+                    <ion-button fill="clear" size="small" (click)="cancelRoomRegistration()">
+                        <ion-icon slot="icon-only" color="danger" name="trash"></ion-icon>
+                    </ion-button>
+                    }
                     <ion-button fill="clear" size="small" (click)="isRegisterRoomOpen = false">
                         <ion-icon slot="icon-only" color="danger" name="close"></ion-icon>
                     </ion-button>
